conformance/bindingTesting/queues/cl_commandQueue_enqueueReadImage.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueReadImage conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

try {
    debug("WebCLCommandQueue :: enqueueReadImage() test case. \n ");
    debug("( WebCLImage image, CLboolean blockingRead, sequence<CLuint> origin, sequence<CLuint> region, CLuint hostRowPitch, ArrayBufferView hostPtr, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var imageWidth = 10;
    var imageHeight = 10;
    var hostRowPitch = 0;
    var origin = [0, 0];
    var region = [imageWidth, imageHeight];

    // MEM_READ_ONLY images.
    var webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, imageWidth, imageHeight);
    var arrayType = wtu.getArrayTypeForChanneltype(webCLImageDescriptorObject.channelType);
    var hostPtr = eval("new " + arrayType + "(imageWidth * imageHeight * wtu.getBytesForChannelOrder(webCLImageDescriptorObject.channelOrder))");
    var webCLImageRead = wtu.createImage(webCLContext, webcl.MEM_READ_ONLY, webCLImageDescriptorObject);
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageRead, true, origin, region, hostRowPitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageRead, false, origin, region, hostRowPitch, hostPtr);");

    // MEM_WRITE_ONLY images.
    webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_WRITE_ONLY, imageWidth, imageHeight);
    arrayType = wtu.getArrayTypeForChanneltype(webCLImageDescriptorObject.channelType);
    hostPtr = eval("new " + arrayType + "(imageWidth * imageHeight * wtu.getBytesForChannelOrder(webCLImageDescriptorObject.channelOrder))");
    var webCLImageWrite = wtu.createImage(webCLContext, webcl.MEM_WRITE_ONLY, webCLImageDescriptorObject);
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageWrite, true, origin, region, hostRowPitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageWrite, false, origin, region, hostRowPitch, hostPtr);");

    // MEM_READ_WRITE images.
    webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, imageWidth, imageHeight);
    arrayType = wtu.getArrayTypeForChanneltype(webCLImageDescriptorObject.channelType);
    hostPtr = eval("new " + arrayType + "(imageWidth * imageHeight * wtu.getBytesForChannelOrder(webCLImageDescriptorObject.channelOrder))");
    var webCLImageReadWrite = wtu.createImage(webCLContext, webcl.MEM_READ_WRITE, webCLImageDescriptorObject);
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr);");

    // Testing with events.
    var webCLEvent = wtu.createEvent();
    webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr, [], webCLEvent);
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr, [webCLEvent]);");
    var webCLUserEvent = wtu.createUserEvent(webCLContext);
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr, [webCLUserEvent]);");
    wtu.setStatus(webCLUserEvent, webcl.COMPLETE);

    webCLEvent = wtu.createEvent();
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [], webCLEvent);");
    webCLEvent = wtu.createEvent();
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr, [], webCLEvent);");

    // Testing for exceptions.
    var releasedEvent = wtu.createUserEvent(webCLContext);
    wtu.release(releasedEvent);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [releasedEvent], webCLEvent);", "INVALID_EVENT_WAIT_LIST");

    webCLUserEvent = wtu.createUserEvent(webCLContext);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [webCLUserEvent]);", "INVALID_EVENT_WAIT_LIST");
    wtu.setStatus(webCLUserEvent, webcl.COMPLETE);
    webCLEvent = wtu.createEvent();
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [webCLEvent]);", "INVALID_EVENT_WAIT_LIST");

    webCLUserEvent = wtu.createUserEvent(webCLContext);
    wtu.setStatus(webCLUserEvent, -1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [webCLUserEvent]);", "EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST");

    webCLEvent = wtu.createEvent();
    wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr, [webCLUserEvent], webCLEvent);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr, [webCLUserEvent], webCLEvent);", "INVALID_EVENT");

    var releasedImage = wtu.createImage(webCLContext, webcl.MEM_READ_WRITE, webCLImageDescriptorObject);
    wtu.release(releasedImage);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(releasedImage, true, origin, region, hostRowPitch, hostPtr);", "INVALID_MEM_OBJECT");

    var originInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, originInvalid, region, hostRowPitch, hostPtr);", "INVALID_VALUE");
    originInvalid = [imageWidth + 1, imageHeight + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, originInvalid, region, hostRowPitch, hostPtr);", "INVALID_VALUE");
    originInvalid = [0];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, originInvalid, region, hostRowPitch, hostPtr);", "INVALID_VALUE");

    var regionInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, regionInvalid, hostRowPitch, hostPtr);", "INVALID_VALUE");
    regionInvalid = [imageWidth + 1, imageHeight + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, regionInvalid, hostRowPitch, hostPtr);", "INVALID_VALUE");
    regionInvalid = [imageWidth];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, true, origin, regionInvalid, hostRowPitch, hostPtr);", "INVALID_VALUE");

    var hostRowPitchValid = imageWidth * 2;
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, origin, region, hostRowPitchValid, hostPtr);");
    var hostRowPitchInvalid = (imageWidth * Float32Array.BYTES_PER_ELEMENT) - 1;
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageRead, true, origin, region, hostRowPitchInvalid, hostPtr);", "INVALID_VALUE");
    hostRowPitchInvalid = hostPtr.BYTES_PER_ELEMENT - 1;
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadImage(webCLImageRead, true, origin, region, hostRowPitchInvalid, hostPtr);", "INVALID_VALUE");

    var hostPtrInvalid = new Float32Array((imageWidth * imageWidth) / Float32Array.BYTES_PER_ELEMENT - 1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtrInvalid);", "INVALID_VALUE");

    var hostPtrValid = new Float32Array(25);
    shouldBeUndefined("webCLCommandQueue.enqueueReadImage(webCLImageReadWrite, false, [1, 1], [5, 5], hostRowPitch, hostPtrValid);");

    // Testing for exception "INVALID_CONTEXT"
    var webCLPlatform = wtu.getPlatform();
    var devices = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_ALL);
    if (devices.length > 1) {
        var webCLDevice1 = devices[0];
        var webCLContext1 = wtu.createContext(webCLDevice1);
        var webCLCommandQueue1 = wtu.createCommandQueue(webCLContext1, webCLDevice1);
        var webCLImageDescriptor1 = wtu.getSupportedImageFormats(webCLContext1, webcl.MEM_READ_WRITE, imageWidth, imageHeight);
        var webCLImage1 = wtu.createImage(webCLContext1, webcl.MEM_READ_WRITE, webCLImageDescriptor1);

        var webCLDevice2 = devices[1];
        var webCLContext2 = wtu.createContext(webCLDevice2);
        var webCLCommandQueue2 = wtu.createCommandQueue(webCLContext2, webCLDevice2);
        var webCLImageDescriptor2 = wtu.getSupportedImageFormats(webCLContext2, webcl.MEM_READ_WRITE, imageWidth, imageHeight);
        var webCLImage2 = wtu.createImage(webCLContext2, webcl.MEM_READ_WRITE, webCLImageDescriptor2);

        shouldThrowExceptionName("webCLCommandQueue1.enqueueReadImage(webCLImage2, true, origin, region, hostRowPitch, hostPtr);", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueue2.enqueueReadImage(webCLImage1, true, origin, region, hostRowPitch, hostPtr);", "INVALID_CONTEXT");

        var webCLUserEvent1 = wtu.createUserEvent(webCLContext1);
        var webCLUserEvent2 = wtu.createUserEvent(webCLContext2);
        shouldThrowExceptionName("webCLCommandQueue1.enqueueReadImage(webCLImage1, true, origin, region, hostRowPitch, hostPtr, [webCLUserEvent2]);", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueue2.enqueueReadImage(webCLImage2, true, origin, region, hostRowPitch, hostPtr, [webCLUserEvent1]);", "INVALID_CONTEXT");
        wtu.setStatus(webCLUserEvent1, webcl.COMPLETE);
        wtu.setStatus(webCLUserEvent2, webcl.COMPLETE);

        var webCLEvent1 = wtu.createEvent();
        wtu.enqueueReadImage(webCLCommandQueue1, webCLImage1, true, origin, region, hostRowPitch, hostPtr, [], webCLEvent1);
        shouldThrowExceptionName("webCLCommandQueue2.enqueueReadImage(webCLImage2, true, origin, region, hostRowPitch, hostPtr, [webCLEvent1]);", "INVALID_CONTEXT");
        var webCLEvent2 = wtu.createEvent();
        wtu.enqueueReadImage(webCLCommandQueue2, webCLImage2, true, origin, region, hostRowPitch, hostPtr, [], webCLEvent2);
        shouldThrowExceptionName("webCLCommandQueue1.enqueueReadImage(webCLImage1, true, origin, region, hostRowPitch, hostPtr, [webCLEvent2]);", "INVALID_CONTEXT");
    }

    // FIXME Need to add test case for INVALID_IMAGE_SIZE.

} catch (e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
